Move scheduler service package
