+++ cmd/start_all.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/dagu-dev/dagu/internal/config"
 	"github.com/dagu-dev/dagu/internal/frontend"
-	"github.com/dagu-dev/dagu/internal/service/scheduler"
+	"github.com/dagu-dev/dagu/internal/scheduler"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"go.uber.org/fx"
