+++ internal/scheduler/entryreader/testdata/scheduled_job.yaml
