+++ internal/scheduler/entryreader/testdata/start_stop.yaml
