+++ internal/scheduler/entryreader/testdata/stop.yaml
